moved in the wrapper classes needed because of Eigen
The issue with Eigen in RTT dataflow is that Eigen requires
the structures to be properly aligned on 8 or 16 byte
boundaries, which we cannot guarantee in the RTT dataflow.

The workaround is to convert back and forth between
Eigen-related types and wrapper types at the task
boundaries. The necessary wrapper types are stored in
this package from now on.